atom-2024-skynet/requirements.txt
fastapi
pillow
tensorflow